+++ packages/babel-helper-wrap-function/.npmignore
@@ -1,0 +1,3 @@
+src
+test
+*.log
