+++ packages/babel-helper-wrap-function/package.json
@@ -1,0 +1,14 @@
+{
+  "name": "babel-helper-wrap-function",
+  "version": "7.0.0-alpha.15",
+  "description": "Helper to wrap functions inside a function call.",
+  "repository": "https://github.com/babel/babel/tree/master/packages/babel-helper-wrap-function",
+  "license": "MIT",
+  "main": "lib/index.js",
+  "dependencies": {
+    "babel-helper-function-name": "7.0.0-alpha.15",
+    "babel-template": "7.0.0-alpha.15",
+    "babel-traverse": "7.0.0-alpha.15",
+    "babel-types": "7.0.0-alpha.15"
+  }
+}
